# service/src/main/java/com/weather/service/dto/WeatherResponse.java
package com.weather.service.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class WeatherResponse {
    private String city;
    private List<DailyForecast> forecasts;
    private String timestamp;
    private boolean offline;
}